requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.2
redis>=4.0
pytest-django>=4.5
python-decouple>=3.6
cloudinary>=1.29